Simulate a Bayesian Bandit choice for a given set of probabilities and a number of points

Simulate a Bayesian Bandit choice for a given set of probabilities and a number of points

ov_simulate_setter_distribution(
  dvw,
  play_phase = c("Reception", "Transition"),
  n_sim = 500,
  priors = list(name = "beta", par1 = 1, par2 = 1),
  epsilon = 1,
  filter_sim = FALSE,
  attack_options = "use_data",
  setter_position_by = "rotation",
  history_table = NULL,
  attack_by = "code",
  exclude_attacks = c("PR"),
  shiny_progress = NULL
)

Arguments

dvw

string or datavolley: a datavolley object as returned by datavolley::dv_read() or a path to datavolley file

play_phase

character: one or both of "Reception", "Transition"

n_sim

integer: number of simulations

priors

numeric: prior distribution of the kill rate for the different attacking options

epsilon

numeric: reward size

filter_sim

logical:

attack_options

string: either "use_data" or "use_history"

setter_position_by

string: either "rotation" or "front_back"

history_table

list: (only if attack_options is "use_history") the object returned by ov_create_history_table()

attack_by

string: either "code", "zone", "tempo", "setter call"

exclude_attacks

character: vector of attack codes to exclude

shiny_progress

numeric: an optional two-element vector. If not NULL or NA, shiny::setProgress() calls will be made during simulation with values in this range

See also

Examples

dvw <- ovdata_example("190301_kats_beds")
system.time({
  ssd <- ov_simulate_setter_distribution(dvw = dvw, play_phase = "Reception",
                                         n_sim = 100, attack_by = "setter call",
                                         setter_position_by = "front_back")
})
#>    user  system elapsed 
#>   2.968   0.012   2.961